adding codebook functions
